Fix warning when running combiner script
Saw this warning: /usr/lib/python3.7/runpy.py:125: RuntimeWarning: 'hail.experimental.vcf_combiner' found in sys.modules after import of package 'hail.experimental', but prior to execution of 'hail.experimental.vcf_combiner'; this may result in unpredictable behaviour

when running:

    python -m hail.experimental.vcf_combiner -h

This restructure fixes it while not changing the export point of
`lgt_to_gt`.
